Split combine.cpp
