+++ src/combine_skip.h
@@ -1,0 +1,80 @@
+/**
+ * @file combine_skip.h
+ *
+ * @author  Guy Maurel
+ * @license GPL v2+
+ * extract from combine.cpp
+ */
+
+#ifndef COMBINE_SKIP_H_INCLUDED
+#define COMBINE_SKIP_H_INCLUDED
+
+#include "chunk_list.h"
+
+
+/**
+ * Skips the D 'align()' statement and the colon, if present.
+ *    align(2) int foo;  -- returns 'int'
+ *    align(4):          -- returns 'int'
+ *    int bar;
+ */
+chunk_t *skip_align(chunk_t *start);
+
+
+/**
+ * Skips everything until a comma or semicolon at the same level.
+ * Returns the semicolon, comma, or close brace/paren or nullptr.
+ */
+chunk_t *skip_expression(chunk_t *start);
+
+
+/**
+ * Skips the list of class/struct parent types.
+ */
+chunk_t *skip_parent_types(chunk_t *colon);
+
+
+/**
+ * Skips over the rest of the template if ang_open is indeed a CT_ANGLE_OPEN.
+ * Points to the chunk after the CT_ANGLE_CLOSE.
+ * If the chunk isn't an CT_ANGLE_OPEN, then it is returned.
+ */
+chunk_t *skip_template_next(chunk_t *ang_open);
+
+
+/**
+ * Skips over the rest of the template if ang_close is indeed a CT_ANGLE_CLOSE.
+ * Points to the chunk before the CT_ANGLE_OPEN
+ * If the chunk isn't an CT_ANGLE_CLOSE, then it is returned.
+ */
+chunk_t *skip_template_prev(chunk_t *ang_close);
+
+
+//! Skips to the start of the next statement.
+chunk_t *skip_to_next_statement(chunk_t *pc);
+
+
+/**
+ * Skips the rest of the array definitions if ary_def is indeed a
+ * CT_TSQUARE or CT_SQUARE_OPEN
+ */
+chunk_t *skip_tsquare_next(chunk_t *ary_def);
+
+
+/**
+ * If attr is CT_ATTRIBUTE, then skip it and the parens and return the chunk
+ * after the CT_FPAREN_CLOSE.
+ * If the chunk isn't an CT_ATTRIBUTE, then it is returned.
+ */
+chunk_t *skip_attribute_next(chunk_t *attr);
+
+
+/**
+ * If fp_close is a CT_FPAREN_CLOSE with a parent of CT_ATTRIBUTE, then skip it
+ * and the '__attribute__' thingy and return the chunk before CT_ATTRIBUTE.
+ * Otherwise return fp_close.
+ */
+chunk_t *skip_attribute_prev(chunk_t *fp_close);
+
+
+#endif /* COMBINE_SKIP_H_INCLUDED */
